Move AA logic to Protocol Kit
