--- packages/account-abstraction-kit/src/constants/index.ts
@@ -1,5 +1,0 @@
-export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
-
-// keccak256(toUtf8Bytes('Safe Account Abstraction'))
-export const PREDETERMINED_SALT_NONCE =
-  '0xb1073742015cbcf5a3a4d9d1ae33ecf619439710b89475f92e2abd2117e90f90'
